add 3.4.1 schema

+++ sdk/src/main/java/org/zstack/sdk/VipBandwidthSpendingDetails.java
@@ -28,11 +28,11 @@ public class VipBandwidthSpendingDetails  {
         return this.spending;
     }
 
-    public double bandwidthSize;
-    public void setBandwidthSize(double bandwidthSize) {
+    public long bandwidthSize;
+    public void setBandwidthSize(long bandwidthSize) {
         this.bandwidthSize = bandwidthSize;
     }
-    public double getBandwidthSize() {
+    public long getBandwidthSize() {
         return this.bandwidthSize;
     }
